requirements.txt
alembic==1.16.5
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.10.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
